Loop if error on secgroup deletion
Sometimes the API returns an error if you try to delete a key
which is 'still in use' for an instance which shows as terminated.

So instead, we check to see if we get an error and then loop until
we don't get an error object.  We compare this non-error to the
expected value of 1.